implement package edit action
